+++ pyproj/transformer.py
@@ -9,8 +9,11 @@ __all__ = [
     "TransformerGroup",
     "AreaOfInterest",
 ]
+import threading
 import warnings
+from abc import ABC, abstractmethod
 from array import array
+from dataclasses import dataclass
 from itertools import chain, islice
 from pathlib import Path
 from typing import Any, Iterable, Iterator, List, Optional, Tuple, Union
@@ -31,11 +34,107 @@ from pyproj.sync import _download_resource_file
 from pyproj.utils import _convertback, _copytobuffer
 
 
+class TransformerMaker(ABC):
+    """
+    .. versionadded:: 3.1
+
+    Base class for generating new instances
+    of the Cython _Transformer class for
+    thread safety in the Transformer class.
+    """
+
+    @abstractmethod
+    def __call__(self) -> _Transformer:
+        """
+        Returns
+        -------
+        _Transformer
+        """
+        raise NotImplementedError
+
+
+@dataclass(frozen=True)
+class TransformerUnsafe(TransformerMaker):
+    """
+    .. versionadded:: 3.1
+
+    Returns the original Cython _Transformer
+    and is not thread-safe.
+    """
+
+    transformer: _Transformer
+
+    def __call__(self) -> _Transformer:
+        """
+        Returns
+        -------
+        _Transformer
+        """
+        return self.transformer
+
+
+@dataclass(frozen=True)
+class TransformerFromCRS(TransformerMaker):
+    """
+    .. versionadded:: 3.1
+
+    Generates a Cython _Transformer class from input CRS data.
+    """
+
+    crs_from: CRS
+    crs_to: CRS
+    skip_equivalent: bool
+    always_xy: bool
+    area_of_interest: Optional[AreaOfInterest]
+    authority: Optional[str]
+    accuracy: Optional[float]
+    allow_ballpark: Optional[bool]
+
+    def __call__(self) -> _Transformer:
+        """
+        Returns
+        -------
+        _Transformer
+        """
+        return _Transformer.from_crs(
+            self.crs_from._crs,
+            self.crs_to._crs,
+            skip_equivalent=self.skip_equivalent,
+            always_xy=self.always_xy,
+            area_of_interest=self.area_of_interest,
+            authority=self.authority,
+            accuracy=self.accuracy,
+            allow_ballpark=self.allow_ballpark,
+        )
+
+
+@dataclass(frozen=True)
+class TransformerFromPipeline(TransformerMaker):
+    """
+    .. versionadded:: 3.1
+
+    Generates a Cython _Transformer class from input pipeline data.
+    """
+
+    proj_pipeline: str
+
+    def __call__(self) -> _Transformer:
+        """
+        Returns
+        -------
+        _Transformer
+        """
+        return _Transformer.from_pipeline(cstrencode(self.proj_pipeline))
+
+
 class TransformerGroup(_TransformerGroup):
     """
     The TransformerGroup is a set of possible transformers from one CRS to another.
 
     .. versionadded:: 2.3.0
+
+    .. warning:: CoordinateOperation and Transformer objects
+                 returned are not thread-safe.
 
     From PROJ docs::
 
@@ -86,7 +185,7 @@ class TransformerGroup(_TransformerGroup):
             area_of_interest=area_of_interest,
         )
         for iii, transformer in enumerate(self._transformers):
-            self._transformers[iii] = Transformer(transformer)
+            self._transformers[iii] = Transformer(TransformerUnsafe(transformer))
 
     @property
     def transformers(self) -> List["Transformer"]:
@@ -168,6 +267,18 @@ class TransformerGroup(_TransformerGroup):
         )
 
 
+class TransformerLocal(threading.local):
+    """
+    Threading local instance for cython _Transformer class.
+
+    For more details, see:
+    https://github.com/pyproj4/pyproj/issues/782
+    """
+
+    def __init__(self):
+        self.transformer = None  # Initialises in each thread
+
+
 class Transformer:
     """
     The Transformer class is for facilitating re-using
@@ -180,14 +291,33 @@ class Transformer:
 
     """
 
-    def __init__(self, base_transformer: Optional[_Transformer] = None) -> None:
-        if not isinstance(base_transformer, _Transformer):
+    def __init__(
+        self,
+        transformer_maker: Union[TransformerMaker, None] = None,
+    ) -> None:
+        if not isinstance(transformer_maker, TransformerMaker):
             ProjError.clear()
             raise ProjError(
                 "Transformer must be initialized using: "
                 "'from_crs', 'from_pipeline', or 'from_proj'."
             )
-        self._transformer = base_transformer
+
+        self._local = TransformerLocal()
+        self._local.transformer = transformer_maker()
+        self._transformer_maker = transformer_maker
+
+    @property
+    def _transformer(self):
+        """
+        The Cython _Transformer object for this thread.
+
+        Returns
+        -------
+        _Transformer
+        """
+        if self._local.transformer is None:
+            self._local.transformer = self._transformer_maker()
+        return self._local.transformer
 
     @property
     def name(self) -> str:
@@ -340,12 +470,16 @@ class Transformer:
         skip_equivalent: bool = False,
         always_xy: bool = False,
         area_of_interest: Optional[AreaOfInterest] = None,
+        authority: Optional[str] = None,
+        accuracy: Optional[float] = None,
+        allow_ballpark: Optional[bool] = None,
     ) -> "Transformer":
         """Make a Transformer from a :obj:`pyproj.crs.CRS` or input used to create one.
 
         .. versionadded:: 2.1.2 skip_equivalent
         .. versionadded:: 2.2.0 always_xy
         .. versionadded:: 2.3.0 area_of_interest
+        .. versionadded:: 3.1.0 authority, accuracy, allow_ballpark
 
         Parameters
         ----------
@@ -363,6 +497,22 @@ class Transformer:
             Default is false.
         area_of_interest: :class:`pyproj.transformer.AreaOfInterest`, optional
             The area of interest to help select the transformation.
+        authority: str, optional
+            When not specified, coordinate operations from any authority will be
+            searched, with the restrictions set in the
+            authority_to_authority_preference database table related to the
+            authority of the source/target CRS themselves. If authority is set
+            to “any”, then coordinate operations from any authority will be
+            searched. If authority is a non-empty string different from "any",
+            then coordinate operations will be searched only in that authority
+            namespace (e.g. EPSG). Requires PROJ 8+.
+        accuracy: float, optional
+            The minimum desired accuracy (in metres) of the candidate
+            coordinate operations. Requires PROJ 8+.
+        allow_ballpark: bool, optional
+            Set to False to disallow the use of Ballpark transformation
+            in the candidate coordinate operations. Default is to allow.
+            Requires PROJ 8+.
 
         Returns
         -------
@@ -370,12 +520,15 @@ class Transformer:
 
         """
         return Transformer(
-            _Transformer.from_crs(
-                CRS.from_user_input(crs_from)._crs,
-                CRS.from_user_input(crs_to)._crs,
+            TransformerFromCRS(
+                CRS.from_user_input(crs_from),
+                CRS.from_user_input(crs_to),
                 skip_equivalent=skip_equivalent,
                 always_xy=always_xy,
                 area_of_interest=area_of_interest,
+                authority=authority,
+                accuracy=accuracy,
+                allow_ballpark=allow_ballpark,
             )
         )
 
@@ -395,7 +548,7 @@ class Transformer:
         Transformer
 
         """
-        return Transformer(_Transformer.from_pipeline(cstrencode(proj_pipeline)))
+        return Transformer(TransformerFromPipeline(proj_pipeline))
 
     def transform(
         self,