+++ test/test_transformer.py
@@ -1,3 +1,4 @@
+import concurrent.futures
 import os
 from functools import partial
 from glob import glob
@@ -1117,3 +1118,74 @@ def test_transformer_group__download_grids__directory(
                 ],
                 any_order=True,
             )
+
+
+@pytest.mark.skipif(
+    pyproj._datadir._USE_GLOBAL_CONTEXT, reason="Global Context not Threadsafe."
+)
+def test_transformer_multithread__pipeline():
+    # https://github.com/pyproj4/pyproj/issues/782
+    trans = Transformer.from_pipeline(
+        "+proj=pipeline +step +inv +proj=cart +ellps=WGS84 "
+        "+step +proj=unitconvert +xy_in=rad +xy_out=deg"
+    )
+
+    def transform(num):
+        return trans.transform(-2704026.010, -4253051.810, 3895878.820)
+
+    with concurrent.futures.ThreadPoolExecutor(max_workers=2) as executor:
+        for result in executor.map(transform, range(10)):
+            pass
+
+
+@pytest.mark.skipif(
+    pyproj._datadir._USE_GLOBAL_CONTEXT, reason="Global Context not Threadsafe."
+)
+def test_transformer_multithread__crs():
+    # https://github.com/pyproj4/pyproj/issues/782
+    trans = Transformer.from_crs(4326, 3857)
+
+    def transform(num):
+        return trans.transform(1, 2)
+
+    with concurrent.futures.ThreadPoolExecutor(max_workers=2) as executor:
+        for result in executor.map(transform, range(10)):
+            pass
+
+
+@pytest.mark.skipif(not PROJ_GTE_8, reason="Requires PROJ 8+")
+def test_transformer_accuracy_filter():
+    with pytest.raises(ProjError):
+        Transformer.from_crs("EPSG:4326", "EPSG:4258", accuracy=0.05)
+
+
+@pytest.mark.skipif(PROJ_GTE_8, reason="Warning for PROJ<8")
+def test_transformer_accuracy_filter_warning():
+    with pytest.warns(UserWarning, match="accuracy requires PROJ 8+"):
+        Transformer.from_crs("EPSG:4326", "EPSG:4258", accuracy=0.05)
+
+
+@pytest.mark.skipif(not PROJ_GTE_8, reason="Requires PROJ 8+")
+def test_transformer_allow_ballpark_filter():
+    with pytest.raises(ProjError):
+        Transformer.from_crs(
+            "EPSG:4326", "EPSG:4258", authority="PROJ", allow_ballpark=False
+        )
+
+
+@pytest.mark.skipif(PROJ_GTE_8, reason="Warning for PROJ<8")
+def test_transformer_allow_ballpark_filter_warning():
+    with pytest.warns(UserWarning, match="allow_ballpark requires PROJ 8+"):
+        Transformer.from_crs("EPSG:4326", "EPSG:4258", allow_ballpark=False)
+
+
+@pytest.mark.skipif(not PROJ_GTE_8, reason="Requires PROJ 8+")
+def test_transformer_authority_filter():
+    transformer = Transformer.from_crs("EPSG:4326", "EPSG:4258", authority="PROJ")
+    assert transformer.description == "Ballpark geographic offset from WGS 84 to ETRS89"
+
+
+@pytest.mark.skipif(PROJ_GTE_8, reason="Warning for PROJ<8")
+def test_transformer_authority_filter_warning():
+    with pytest.warns(UserWarning, match="authority requires PROJ 8+"):
+        Transformer.from_crs("EPSG:4326", "EPSG:4258", authority="PROJ")